+++ lib/uti/my_network/BlockingMultiThread/IClientUdpMultiThreadWrapper.hpp
@@ -13,7 +13,7 @@
 namespace uti::network {
     class IClientUdpMultiThreadWrapper {
         protected:
-            virtual void setServer(const std::string &serverAddress, size_t port) = 0;
+            virtual void setServer(const std::string &serverAddress, unsigned int port) = 0;
             virtual void sendMessage(const boost::any & message) = 0;
             virtual std::string getReply() = 0;
             virtual void stop() = 0;
